requirements.txt
asgiref==3.7.2
Django==3.2
django-ckeditor==6.7.0
django-cleanup==8.0.0
django-countries==7.5.1
django-js-asset==2.2.0
django-modeltranslation==0.18.11
django-mptt==0.16.0
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
phonenumbers==8.13.28
phonenumberslite==8.13.28
pillow==10.2.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
